refactor: move server/files tests into test/server/files
